# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STRESSNES luxury fashion ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized

  # ── Products ────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all active categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── Collections ─────────────────────────────────────────────
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get the current cart
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    post:
      operationId: addToCart
      tags: [cart]
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear all items from the cart
      responses:
        "200":
          description: Cart cleared
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update a cart item quantity
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove an item from the cart
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item removed

  # ── Orders ──────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, fullName, role]
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        token:
          type: string

    # ── Products ───────────────────────────────────────────────
    ProductImage:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string
        altText:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
    ProductVariant:
      type: object
      required: [id, sku, isActive]
      properties:
        id:
          type: string
        sku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        material:
          type: ["string", "null"]
        priceOverride:
          type: ["number", "null"]
        isActive:
          type: boolean
        stockQty:
          type: integer
    ProductCategoryRef:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    Product:
      type: object
      required: [id, title, slug, sku, price, status]
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        status:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        category:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        collection:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        createdAt:
          type: string
          format: date-time
...
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer

    # ── Collections ────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: string
        productId:
          type: string
        variantId:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        product:
          $ref: "#/components/schemas/Product"
    Cart:
      type: object
      required: [id, items]
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        variantId:
          type: string
        quantity:
          type: integer
          minimum: 1
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1

    # ── Orders ─────────────────────────────────────────────────
...