# lib/api-zod/src/generated/types/productCategoryRef.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * STRESSNES luxury fashion ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCategoryRef {
  id: string;
  name: string;
  slug: string;
}